Rename "aliase" package to "aliases" for consistency
